Move leaderboard code into substitution.py
