+++ substitution.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from flask import *
 from database import database
+from datetime import datetime, tzinfo
 import random
 import json
 import re
@@ -89,3 +90,28 @@ def play(difficulty, idx=-1, language=None):
         foreign=len(foreign[lang].intersection(text.upper())) > 0
     )
 
+@app.route("/leaderboard_insert", methods=['POST'])
+def leaderboard_insert():
+    name = request.form['name']
+    time = int(request.form['time_solved'])
+    print time
+    time_solved = datetime.utcfromtimestamp(time)
+    print time_solved
+    difficulty = request.form['difficulty']
+
+    db = database.dbcon()
+    cur = db.cursor()
+    query = 'INSERT INTO crypto_leaderboard (name, difficulty, time_solved) VALUES (%s, %s, %s)'
+    cur.execute(query, (name, difficulty, time_solved))
+    cur.execute('COMMIT')
+    cur.close()
+    return json.dumps({'status': 'OK'})
+
+@app.route("/leaderboard")
+def leaderboard():
+    db = database.dbcon()
+    cur = db.cursor()
+    query = 'SELECT * FROM crypto_leaderboard'
+    cur.execute(query)
+    users = [[x[1], x[2], x[3].strftime('%H:%M:%S')] for x in cur.fetchall()]
+    return render_template("substitution.leaderboard.html", users=users)
