+++ templates/main.html
@@ -46,7 +46,7 @@
                                 <li class="divider"></li>
                                 <li><a href="{{url_for('substitution.play', difficulty='caesar')}}">Cezarjeva šifra</a></li>
                                 <li><a href="{{url_for('substitution.play', difficulty='ready')}}">Pripravljene zamenjalke</a></li>
-                                <li><a href="{{url_for('leaderboard_get')}}">Lestvica</a></li>
+                                <li><a href="{{url_for('substitution.leaderboard')}}">Lestvica</a></li>
                               </ul>
                             </li>
                         </ul>
